optimized upgrade values

+++ GMD Unity Project/Assets/Scripts/UpgradeManager/ResetUpgrades.cs
@@ -26,7 +26,7 @@ public class ResetUpgrades : MonoBehaviour
     public void ResetUpgrade()
     {
         PlayerPrefs.SetInt(IUpgradeables.coinCountKey, 9999);
-        PlayerPrefs.SetInt(IUpgradeables.healthKey, 1);
+        PlayerPrefs.SetFloat(IUpgradeables.healthKey, 1);
         PlayerPrefs.SetFloat(IUpgradeables.movementSpeedKey, 3);
         PlayerPrefs.SetFloat(IUpgradeables.waterSpeedKey, 2);
         PlayerPrefs.SetFloat(IUpgradeables.jumpForceKey, 5);
